requirements.txt
numpy==1.21.0
pandas==1.2.5
Pillow==8.2.0
PySimpleGUI==4.45.0
matplotlib==3.4.2
PyYAML==5.4.1
py7zr==0.16.1
tables==3.6.1
